Unify branding across web and mini app
